2018-11-22-Buergerpreis.md: use jekyll-image-tag plugin to generate smaller image file
Reduce resolution to 800 pixel width (ca. twice the size that will be used
by browser at 100% and 96 dpi). This gives a good image even when zooming
in, and should also give good results on a high-DPI screen.

The complicated "width:30%;min-width:400px" style has been removed, as it
caused the image to be simply displayed at 400 pixel width all the time.